+++ rummage/lib/gui/settings/__init__.py
@@ -963,7 +963,9 @@ class Settings:
         notify.setup_notifications(
             "Rummage",
             img,
-            (notifier, None)
+            term_notify=notifier,
+            sender=None,
+            sound=os.path.join(data.RESOURCE_PATH, "glass-ding.wav")
         )
         notify.setup_growl_notifications("Rummage", growl_png)
         notify.enable_growl(cls.get_notify_method() == "growl" and notify.has_growl())