+++ docs/src/markdown/changelog.md
@@ -9,6 +9,8 @@
 - **FIX**: Ensure column sizing includes header size.
 - **FIX**: When showing timed status event for results in status bar (full path name), decrease length of time that the temporary status is shown and clear temporary status when the cursor leaves result items.
 - **FIX**: Don't log handled notification errors when a notification system is not found as these are handled gracefully and there is no need to worry the user.
+- **FIX**: Include a default notification sound and use it as the default across all platforms.
+- **FIX**: Provide fallback players for Linux: `paplay`, `aplay`, and `play` (`sox`).
 
 ## 4.6.3